revamped figures
It's a good thing that this is a solo project. I may have gotten carried away on several tangents and neglected good commit habbits.
This started to feel like driving down a mountain when your brakes have failed and you're just hoping for the best. I think it worked
this time, but maybe not the best feeling.